requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
gunicorn>=21.2.0

